add new testdata instead of changing
