+++ gotests_test.go
@@ -66,7 +66,7 @@ func TestGenerateTests(t *testing.T) {
 		}, {
 			name: "Target test file",
 			args: args{
-				srcPath:  `testdata/test100_test.go`,
+				srcPath:  `testdata/test103_test.go`,
 				only:     regexp.MustCompile("wrapToString"),
 				subtests: true,
 			},
