+++ testdata/goldens/existing_test_file.go
@@ -1,7 +1,6 @@
 package testdata
 
 import (
-	"fmt"
 	"reflect"
 	"testing"
 )
@@ -44,14 +43,6 @@ func TestBaz100(t *testing.T) {
 			t.Errorf("%q. baz100() = %v, want %v", tt.name, got, tt.want)
 		}
 	}
-}
-
-func wrapToString(in []int) []string {
-	var result []string
-	for _, x := range in {
-		result = append(result, fmt.Sprintf("%v", x))
-	}
-	return result
 }
 
 func TestFoo100(t *testing.T) {
